backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
aiomysql==0.2.0
sqlalchemy==2.0.23
httpx==0.25.1
email-validator==2.1.0
